fix: change Account monetary fields from float to int
Banking systems should use integer amounts for precision. This ensures
consistency with:
- Database schema (INTEGER type in original teabank.py)
- Transaction model (uses int)
- Banking best practices (avoid floating-point precision issues)

Changed fields: amount, pending, share, total_balance return type

+++ src/models/account.py
@@ -10,11 +10,11 @@ class Account:
     id: int
     account_no: str
     name: str
-    amount: float
-    pending: float
-    share: float
+    amount: int
+    pending: int
+    share: int
 
     @property
-    def total_balance(self) -> float:
+    def total_balance(self) -> int:
         """Calculate total balance including pending amounts."""
         return self.amount + self.pending
